\name{lmtestrob-package}
\alias{lmtestrob-package}
%\alias{lmtestrob}
\docType{package}
\title{
Outlier Robust Specification Testing
}
\description{
The package contains an outlier robust functional miss-specification test.
}
\details{
\tabular{ll}{
Package: \tab lmtestrob \cr
Type: \tab Package \cr
Version: \tab 0.1 \cr
Date: \tab 2023-05-22 \cr
License: \tab GPL-2 \cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Mikhail Zhelonkin (aut, cre).

Maintainer: Mikhail Zhelonkin <Mikhail.Zhelonkin@gmail.com>
}
\references{
Heritier, S., and Ronchetti, E. (1994) Robust Bounded-Influence Tests in General Parametric Models. \emph{Journal of the American Statistical Association}, 89, p. 897-904.

Zeileis, A., and Hothorn, T. (2002) Diagnostic Checking in Regression Relationships. \emph{R News}, 2, p. 7-10.
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\seealso{
\code{\link{robfmtest}}
}
\keyword{ Robust inference }
%\examples{
%~~ simple examples of the most important functions ~~
%}
