\name{print.robfmtest}
\alias{print.robfmtest}
\title{
Print a \code{robfmtest} Object
}
\description{
Print an object generated by \code{\link{robfmtest}}
}
\usage{
\method{print}{robfmtest}(x, digits = 4, ...)
}
\arguments{
  \item{x}{
object returned from the \code{\link{robfmtest}}. 
}
  \item{digits}{
number of significant digits to be printed.
}
  \item{\dots}{
currently not used.
}
}
\value{
No return value.
}
\author{
Mikhail Zhelonkin
}
\seealso{
\code{\link{robfmtest}}
}
